+++ classes/Report.class.php
@@ -762,13 +762,6 @@ class Report
                 $dt->format($_SHOP_CONF['datetime_fmt'], true) . '</span>';
             break;
 
-        case 'order_total':
-            $sales = SHOP_getVar($A, 'sales_amt', 'float');
-            $tax = SHOP_getVar($A, 'tax', 'float');
-            $shipping = SHOP_getVar($A, 'shipping', 'float');
-            $retval = Currency::getInstance($A['currency'])->FormatValue($sales + $tax + $shipping);
-            break;
-
         case 'status':
             // Show the order status. Admins can update the status, Users can view only.
             if ($extra['isAdmin']) {
@@ -814,10 +807,14 @@ class Report
             }
             break;
 
+        case 'order_total':
         case 'total':
         case 'shipping':
+        case 'handling':
+        case 'net_taxable':
+        case 'net_nontax':
         case 'tax':
-            $retval = Currency::getInstance($A['currency'])->FormatValue($fieldvalue);
+            $retval = Currency::getInstance($A['currency'])->FormatValue((float)$fieldvalue);
             break;
 
         case 'customer':