+++ language/english_utf-8.php
@@ -85,6 +85,7 @@ $LANG_SHOP = array (
 'disabled'          => 'Disabled',
 'featured'          => 'Featured',
 'taxable'           => 'Taxable',
+'nontaxable'        => 'Non-Taxable',
 'delete'            => 'Delete',
 'thanks_title'      => 'Thank you for your order!',
 'yes'               => 'Yes',
@@ -348,6 +349,10 @@ $LANG_SHOP = array (
     'reorder' => array(
         'name' => 'Item Reorder Report',
         'dscp' => 'List all items at or below their minimum stock levels.',
+    ),
+    'tax' => array(
+        'name' => 'Sales Tax Report',
+        'dscp' => 'List orders including tax and region information',
     ),
 ),
 'my_orders' => 'My Orders',
@@ -663,6 +668,7 @@ $LANG_SHOP = array (
 'zone_rule' => 'Zone Rule',
 'restrictions' => 'Restrictions',
 'none_defined' => 'No %s have been defined.',
+'incl_nontax' => 'Include non-taxable sales?',
 );
 if (isset($_SHOP_CONF['ena_ratings']) && $_SHOP_CONF['ena_ratings']) {
     $LANG_SHOP['list_sort_options']['top_rated'] = 'Top Rated';