+++ language/german.php
@@ -647,6 +647,11 @@ $LANG_SHOP = array (
 'on_sale' => 'On Sale!',
 'chk_all' => 'Check All',
 'default' => 'Default',
+'nontaxable' => 'Non-Taxable',
+'err_dup_iso' => 'An error occurred. Check that the ISO code is unique.',
+'cust_text_override' => 'Custom Text Override',
+'prod_features' => 'Product Features',
+'incl_nontax' => 'Include non-taxable sales?',
 );
 if (isset($_SHOP_CONF['ena_ratings']) && $_SHOP_CONF['ena_ratings']) {
     $LANG_SHOP['list_sort_options']['top_rated'] = 'Top Rated';
@@ -677,6 +682,7 @@ $LANG_SHOP_HELP = array(
 'hlp_var_bulk_edit' => 'Editing multiple product variants at once. Any updates to the below fields will be applied to each selected variant.',
 'hlp_var_after_item' => 'Variants must be added after the item is created.',
 'hlp_bulk_cats' => 'If any categories are selected, all product categories will be deleted and replaced by the new selections.',
+'hlp_prod_feat_form' => 'Select features to add to the product. If a custom text string is entered then it will be used, otherwise the selection from the Feature Value list is used. Check an existing feature to delete it, or add new features by clicking the (+) icon.',
 );
 
 $LANG_MYACCOUNT['pe_shop'] = 'Shopping';